requirements.txt
asyncio

pyyaml
pytz
tqdm
jsonpickle
jinja2
python-dotenv

feedparser
feedgen
feedgenerator

aiohttp
bs4
requests
html2text
urllib3

pillow
numpy<2
opencv-python
pytesseract

mkdocs-material